Re-organized
Reference_model.py is the top file
ALU.py contains ALU design
Modified Final_Assembler.py, Register.py, Shifter.py
New txt files - dump.txt(for general register dump)
              - in25.txt(assembler i/p file for 25th instruction test)
