--- Mem files/in.txt
@@ -1,1 +1,0 @@
-R0 = R1 + R2
