--- Mem files/opcode.txt
@@ -1,1 +1,0 @@
-01000000000000000000001001000000
