+++ Scripts/Register.py
@@ -1,4 +1,4 @@
-import time
+
 reg_bank={
     "R":{"0000":"0000000000000010","0001":"0111000010101010","0010":"0000000000010000","0011":"1111111011111101","0100":"XXXX","0101":"XXXX","0110":"XXXX","0111":"XXXX","1000":"XXXX","1001":"XXXX","1010":"XXXX","1011":"XXXX","1100":"XXXX","1101":"XXXX","1110":"XXXX","1111":"XXXX"},
     "I":{"0000":"XXXX","0001":"XXXX","0010":"XXXX","0011":"XXXX","0100":"XXXX","0101":"XXXX","0110":"XXXX","0111":"XXXX","1000":"XXXX","1001":"XXXX","1010":"XXXX","1011":"XXXX","1100":"XXXX","1101":"XXXX","1110":"XXXX","1111":"XXXX"},
@@ -9,10 +9,10 @@ reg_bank={
 reg={
     "0000":"R",
     "0001":"I",
-    "0010":"M",
+    "0010":"M", 
     "0110":"0110",
     "0111":"0111"
-}
+}    
 reg_name={
     "FADDR":"01100000",
     "DADDR":"01100001",
@@ -21,30 +21,40 @@ reg_name={
     "PCSTKP":"01100101",
     "MODE1":"01111011",
     "ASTAT":"01111100",
-    "STKY":"01111110"}
+    "STKY":"01111110"
+    }
+
 def hex_to_ubin(hnum):
     bnum = format(int(hnum,16),'016b')
     return bnum
+
 def ubin_to_hex(bnum):
     hnum=format(int(bnum,2),'04X')
     return hnum
+
 def int_to_hex(num):
     hnum=format(num,"04X")
     return hnum
+
 def compliment(num):
     s=-2**15
     for i in range(15):
         s=s+(int(num[15-i])*2**i)
     return(-s)
+
 def compliment40(num):
     s=-2**39
     for i in range(39):
         s=s+(int(num[39-i])*2**i)
     return(-s)
+
 def put_to_reg(inst,val):
     reg_bank[reg[inst[0:4]]][inst[4:]]=val
-    #reg_bank.get(reg.get(inst[0:4]))[dec(inst[4:])]=val
 
 def get_from_reg(inst):
     return reg_bank[reg[inst[0:4]]][inst[4:]]
-    #return(reg_bank.get(reg.get(inst[0:4]))[dec(inst[4:])])
+
+def bit_wr(ureg,bit,val):                   # Bit writing function
+    l = list(reg_bank['0111'][ureg])
+    l[15-bit] = val
+    reg_bank['0111'][ureg] = ''.join(l)
