+++ dump.txt
@@ -1,19 +1,19 @@
-R0 : 0000000000000010
-R1 : 0111000010101010
-R2 : 0000000000010000
-R3 : 1111111011111101
-R4 : XXXX
-R5 : XXXX
-R6 : XXXX
-R7 : XXXX
-R8 : XXXX
-R9 : XXXX
-R10 : XXXX
-R11 : XXXX
-R12 : XXXX
-R13 : XXXX
-R14 : XXXX
-R15 : XXXX
+R0 : 0000000000000011
+R1 : 0000000000000001
+R2 : 0000000000000010
+R3 : 0000000000000011
+R4 : 0000000000000100
+R5 : 0000000000000101
+R6 : 0000000000000110
+R7 : 0000000000000111
+R8 : 0000000000001000
+R9 : 0000000000001001
+R10 : 0000000000001010
+R11 : 0000000000001011
+R12 : 0000000000001100
+R13 : 0000000000001101
+R14 : 0000000000001110
+R15 : 0000000000001111
 I0 : XXXX
 I1 : XXXX
 I2 : XXXX
