+++ in25.txt
@@ -1,0 +1,17 @@
+R0 = 0000
+R1 = 0001
+R2 = 0002
+R3 = 0003
+R4 = 0004
+R5 = 0005
+R6 = 0006
+R7 = 0007
+R8 = 0008
+R9 = 0009
+R10 = 000A
+R11 = 000B
+R12 = 000C
+R13 = 000D
+R14 = 000E
+R15 = 000F
+R0 = R1 + R2
