+++ opcode.txt
@@ -14,3 +14,4 @@
 00001100000011010000000000001101
 00001100000011100000000000001110
 00001100000011110000000000001111
+01000000000000000000001001000000
